requirements.txt
Flask
torch==2.5.1 
transformers
gspread
google-auth
